feat: stable diffusion 3 support infer boost algorithm and some base optimization

+++ mindone/diffusers/models/normalization.py
@@ -290,7 +290,7 @@ class LayerNorm(nn.Cell):
         self.layer_norm = ops.LayerNorm(-1, -1, epsilon=eps)
 
     def construct(self, x: Tensor):
-        x, _, _ = self.layer_norm(x, self.weight, self.bias)
+        x, _, _ = self.layer_norm(x, self.weight.to(x.dtype), self.bias.to(x.dtype))
         return x
